+++ frontend/src/ui/controls/position-picker/utils.ts
@@ -1,0 +1,13 @@
+import { ChannelType } from '@vlight/controls'
+
+export const positionChannels: string[] = [
+  ChannelType.Pan,
+  ChannelType.PanFine,
+  ChannelType.Tilt,
+  ChannelType.TiltFine,
+]
+
+export interface Position {
+  [ChannelType.Pan]: number
+  [ChannelType.Tilt]: number
+}